python/oci-{{cookiecutter.project_domain | lower | replace(' ', '-') | replace('_', '-')}}-mcp-server/pyproject.toml
[project]
name = "oracle.oci-{{cookiecutter.project_domain | lower | replace(' ', '-') | replace('_', '-')}}-mcp-server"

version = "0.0.0"
description = "{{cookiecutter.description}}"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp==2.12.2",
    "oci==2.160.0",
    "pydantic==2.12.3",
]
license = "UPL-1.0"
license-files = ["LICENSE.txt"]
authors = [
    {name = "Oracle MCP", email = "237432095+oracle-mcp@users.noreply.github.com"},
]

classifiers = [
    "License :: OSI Approved :: Universal Permissive License (UPL)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
"oracle.oci-{{cookiecutter.project_domain | lower | replace(' ', '-') | replace('_', '-')}}-mcp-server" = "oracle.oci-{{cookiecutter.project_domain | lower | replace(' ', '-') | replace('_', '-') | replace('-', '_')}}_mcp_server.server:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.coverage.run]
omit = [
    "**/__init__.py",
    "**/tests/*",
    "dist/*",
    ".venv/*",
]

[tool.coverage.report]
precision = 2
fail_under = 90
